Move 'MediaCachePolicy' to its own file
