+++ MovieListFramework/Media Cache/MediaCachePolicy.swift
@@ -1,0 +1,24 @@
+//
+//  MediaCachePolicy.swift
+//  MovieListFramework
+//
+//  Created by macbook on 14/09/2023.
+//
+
+import Foundation
+
+internal final class MediaCachePolicy {
+    private init() {}
+    private static let calendar = Calendar(identifier: .gregorian)
+
+    private static var maxCacheAgeInDays: Int {
+        return 7
+    }
+    
+    internal static func validate(_ timestamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
+            return false
+        }
+        return date < maxCacheAge
+    }
+}
